Update to snes-20100809 release.
Code cleanups, no binaries. I'll leave the previous WIP up so you can
get Windows binaries for now.

This adds some sync.sh improvements to make it handle errors more
gracefully.
It also updates asnes a good bit. All of the four base processors now
have all publicly accessible functions right at the top of the main
headers, and everything else is private. This is to allow these
headers to essentially take the place of the previous base classes in
the old bsnes-merged format. So if there's something public there, you
need to implement that exact function to make your own module.
I removed the frame counter from the PPU, as it has nothing to do with
emulation. That now resides inside the Qt -> SNES interface code.
Quite amazing, I was actually saving the frame counter into the save
state files before, yuck.
Removed some baggage in the System class: it was friending a bunch of
long-dead functions and classes.
Forgot to re-add the CHEAT_SYSTEM define to info.hpp, so that's been
put back.

Need to think of something to do about the documentation and about
screen, kind of silly for asnes to display bsnes in certain areas.
I'll also play around with tar --transform sometime tomorrow, no
promises on whether or not I'll use it.