Update to bsnes v067r26 release.
Okay, I'm pretty much ready to release the next version, so heavy
testing on this would be appreciated. To test the debugger (built with
the compatibility core for moderate speed and PPU debugging support),
rename the DLL to an EXE.

I think I'd rather play it safe though and go with v068. While the
performance core is definitely major, I only have one shot at v1.0. It
would at least be nice to have perfect mapping and the BGnxOFS caching
for SMW1 in there first.

Anyway, changelog:

Re-added blargg's DSP, but this time I merged only spc_dsp and also
fixed the initial regs for Dual Orb II. Which restores the 5-10%
speedup on the compatibility and performance cores. Fixed the initial
register values for the fast CPU core, which fixes Armored Police and
the other Atlus game's title screen in the performance core. Added a
missing debugvirtual prefix to op_step, which fixes CPU stepping in
the performance core. I was using the description field for profile
identification in savestates, but that was of course the description
for the state manager. Whoops. Added a new field for profile name to
the save states, and fixed the state manager to work with that change.
Adjusted the about screen colors, which is how you can tell which core
you're using without it being annoying. Probably did some other stuff
too, meh.